Move show_window() from package core to exec script
This moves the interactive visualization window from the package module
and into the executable program that is not included in the Python
package.
The new version of show_window is slightly modified to enable
integration with the config app (to change threshold values during
runtime, etc.)
